requirements.txt
flask==2.3.3
flask-socketio==5.3.6
opencv-python==4.8.1.78
ultralytics==8.0.196
numpy==1.24.3
pillow==10.0.1
requests==2.31.0
python-socketio==5.9.0
eventlet==0.33.3
dashscope==1.14.1
psutil==5.9.5